julia: handle floor for python compat

+++ tests/expected/comb_sort.jl
@@ -4,7 +4,7 @@ function comb_sort(seq::Array{Int64})::Array{Int64}
     gap = length(seq)
     swap = true
     while gap > 1 || swap
-        gap = max(1, floor((gap / 1.25)))
+        gap = max(1, Int64(floor((gap / 1.25))))
         swap = false
         for i in (0:(length(seq)-gap)-1)
             if seq[i+1] > seq[(i+gap)+1]
